Update TermSheet.js
Please note that: 

(i) Have not reviewed the part underneath Clause 16.
(ii) It is currently unclear which part is Schedule A, Schedule B and Schedule C.
(iii) Currently, there is no mention about the PREPAYMENT. 
(iv) all the CONDITIONS PRECEDENT have missed out - think you have removed them on purpose. In such case, you would have to communicate well with the Borrower in advance as to what sort of Conditions Precedent need to be satisfied PRIOR TO the signing of this Smart Contract.